+++ src/main/java/api/dao/IReservaDAO.java
@@ -3,6 +3,6 @@ package api.dao;
 import org.springframework.data.jpa.repository.JpaRepository;
 import api.dto.Reserva;
 
-public interface IReservaDAO extends JpaRepository<Reserva, Integer>{
+public interface IReservaDAO extends JpaRepository<Reserva, Long>{
 
 }
